Disable Bitcode on iOS and macOS
gmp fails to build with bitcode enabled because it incorporates a lot of
assembly code. If we need to support bitcode in the future (tvOS,
watchOS), then we will need to disable assembly for those platforms.

It is much preferred to have the hand-tuned assembly over the bitcode.
Right now XKCP and secp256k1 use generic C implementations, but
ultimately we will want to switch them to their hand-tuned assembly as
well.