Add logging support with the KotlinLogging library
This commit adds support for the microutils KotlinLogging library. It
uses a custom fork that supports Apple's os_log, and will do so until
it is upstreamed into the main repository.

This commit also deletes the OSLogLogger for Koin, and implements a
KotlinLogging logger for Koin. This commit integrates KotlinLogging with
Koin for application code as well. You can ask Koin for a logger with a
given specifier, and it will make an appropriately named logger. The
commit also adds basic logging to AccountStore, logging when an account
is added.

This also adds support for Android with slf4j-android, which serves as a
backed for the slf4j-api library that KotlinLogging uses on Android.