Split up platform module into smaller modules
The idea of the platform module was to put things with platform-specific
implementations in there, but that does not make a lot of sense because
many, many things are goig to have platform-specific implementations. We
should instead organize code around its functionality.

This commit also renames usages of the variable name `bytes` in favor of
the name `data` so that the entire codebase is consistent. I like `data`
because it is consistent with the Data type in Swift.