Implement UI test for import phrase flow
This commit first fixes test running for UI tests on macOS and iOS by
properly embedding the required frameworks. It also annotates the
necessary views with accessibility identifiers, and implements a simple
test that clicks through the import flow. In order to make the test
repeatable, I added support for a "--reset" launch argument to the iOS
app in order to reset the account store.