Call functions on contracts with native functions
This is key for indexing, because I need to interact with the current
state of the contracts to get metadata, and the way to do that is to
call `view` functions with `eth_call`.

In the future, I would like to add code generation based on ABI JSON
files, but that is not needed yet.