Upgrade to Kotlin 1.8.0-RC
This commit also contains a huge refactoring of the Gradle
configuration. The Gradle configuration before this commit used a bunch
of old ways to do things, which ended up being confusing and hard to
manage. I refactored the entire configuration to use a single "version
catalog", which is a feature that landed in Gradle 7.4. I also opted for
the `dependencyResolutionManagement` feature, which centralizes
dependency resolution across the project.

In addition to updating Kotlin, I also updated Gradle to 7.6 and Android
Gradle Plugin to 7.4. I had to upgrade Android Gradle Plugin because
7.3.1 had a compiler error with Kotlin 1.8.0. I also updated a handful
  of dependencies.