Refactor browser message sending architecture
This commit does a number of things. First, it introduces the
BrowserFrameSession object, which is intended to exist for the lifetime
of a specific tab in the browser. This allows separate state to exist
per-tab, which is needed to implement things like sending a `connect`
event on first connect.

I also refactored all of the messages on the extension side to have
type-safe parsing. This is key to adding more messages over time in a
scalable way.

As a part of this refactor, I also moved identifying characteristics
about a session to the explicit `Session` object. This is a much cleaner
abstraction than concatenating strings to form a unique identifier.

With this new architecture, I made it so that the page emits a `connect`
event after `window.ethereum` is accessed for the first time. This is
the next big step towards supporting the full Ethereum provider API