Implement browser-specific extension packaging
This change builds separate versions of the extension for Chrome, Safari
and Firefox. It also switches all compiled javascript to ES modules and
enables code splitting between all of the entrypoints.

I switched to a non-persistent background page in both Firefox and
Safari, while I use a service worker in Chrome. The content script is a
IIFE shim that uses `import()` to bring in the content page ES module.

The build script handles making browser-specific changes to the
manifest.json