Fix deserializing JsonRpcResponse on macOS + iOS
Kotlin Serialization is supposed to be able to find the serializer for a
class at runtime using a type annotation. That doesn't seem to work on
Apple platforms as compiled in this project, so each specific
serialization use case is handled with a custom Json object with a
custom serializersModule. I needed to add JsonRpcResponse to the
serializers module, and that fixed the crash within the serialization
logic.